// src/api-client/model/permissionRequest.ts
/**
 * OpenApi specification - BZDATA
 *
 * Contact: astairenazaire@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PermissionRequest { 
    nomPermission?: string;
    descriptionPermission?: string;
    module?: string;
    canRead?: boolean;
    canWrite?: boolean;
    canEdit?: boolean;
    canDelete?: boolean;
}