// src/api-client/model/roleResponse.ts
/**
 * OpenApi specification - BZDATA
 *
 * Contact: astairenazaire@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PermissionResponse } from './permissionResponse';


export interface RoleResponse { 
    id?: number;
    nomRole?: string;
    descriptionRole?: string;
    permissions?: Set<PermissionResponse>;
    createdDate?: string;
    lastModifiedDate?: string;
}